+++ app/src/main/kotlin/com/getstrm/pace/plugins/Plugin.kt
@@ -1,0 +1,6 @@
+package com.getstrm.pace.plugins
+
+interface Plugin {
+
+    val id: String
+}
