+++ gradle.properties
@@ -7,4 +7,4 @@ jooqVersion=3.18.7
 
 dockertag = pace-local
 
-generatedBufDependencyVersion=00000000000000.59c2e7116922
+generatedBufDependencyVersion=00000000000000.bc3bdf7fade6
