+++ protos/getstrm/pace/api/processing_platforms/v1alpha/processing_platforms.proto
@@ -6,7 +6,7 @@ import "google/api/annotations.proto";
 import "buf/validate/validate.proto";
 import "google/rpc/error_details.proto";
 
-option go_package = "github.com/getstrm/pace/protos/processing_platforms;processing_platforms";
+option go_package = "github.com/getstrm/pace/protos/api/processing_platforms;processing_platforms";
 option java_package = "com.getstrm.pace.api.processing_platforms.v1alpha";
 option java_multiple_files = true;
